build out FilterBar component
