+++ .env.example.test
@@ -12,3 +12,12 @@ PARTICIPANT_PASSWORD='your-participant-password'
 PARTICIPANT_UID='your-participant-uid'
 PARTICIPANT_EMAIL='your-participant-email'
 PARTICIPANT_EMAIL_PASSWORD='your-participant-email-password'
+
+# Parent information for testing
+PARENT_FIRST_NAME='your-parent-first-name'
+PARENT_LAST_NAME='your-parent-last-name'
+PARENT_EMAIL='your-parent-email'
+PARENT_PASSWORD='your-parent-password'
+
+# Firebase App Check Tokens
+VITE_APPCHECK_DEBUG_TOKEN='your-firebase-app-check-debug-token'
