+++ src/components/RoarDataTable.vue
@@ -293,6 +293,7 @@
                     <PvTriStateCheckbox v-model="filterModel.value" input-id="booleanFilter" style="padding-top: 2px" />
                     <label for="booleanFilter">{{ col.header + '?' }}</label>
                   </div>
+
                   <div v-if="col.dataType === 'score'">
                     <PvDropdown
                       v-model="filterModel.value"
@@ -751,7 +752,7 @@ const refFilters = ref(computedFilters.value.computedFilters);
 
 const resetFilters = () => {
   refFilters.value = computedFilters.value.computedFilters;
-  emit('reset-filters');
+  // emit('reset-filters');
 };
 
 let toolTipByHeader = (header) => {
